Implements very basic idea of join() function
Adds simple if statement to join() which checks whether the joining thread (A) is finished, and if not, put the current thread (B) to sleep.
Still testing how to check if thread A has already been joined. Thinking of just creating a new private class boolean to determine whether this thread has already been joined or not.
Also need to figure out how to get thread B to run after thread A has finished. Unsure if that happens automatically, or if I need to add my own code to make sure it happens.
Adds test function joinTest1() used (obviously) for testing purposes. The functionality of joinTest1() will change as needed.
Spacing was modified for readability.
THIS CODE ERRORS, DO NOT USE.