Improves on join() among other functions.
A new KThread private variable was added called joinParent, which is used to reference the "parent" thread (B) of a joined thread (A). The value is set to null when the thread hasn't been joined.
join() now uses joinParent to determine if a thread A has already been joined by another thread B.
run() now checks if a thread's status is either statusBlocked or statusFinished and executes if the thread is not in either of these states.
runThread() has been edited to change the status of a parent thread. The function assumes that after target.run(); the thread has been run fully, and therefore only needs to be finished, using finish();. Inbetween the run() and finish() was an if statement to make a parent thread's status changed back to statusReady.
Very hesitant about runThread(). Will be testing extensively to see how well it works.
Code does not error.