Adds some necessary variables and starts join
Adds 6 new instances variables to class UserProcess:
pid: the process ID number.
status: the exit status of the process.
children: a LinkedList of child processes to the root process.
parent: a reference to the parent process, will stay null until it's evident that the process is a child.
pidTracker: a process ID tracker to make sure every process can have a unique pid.
thread: a thread that the process will need to execute.
join() currently just checks if the passed pid argument is actually a child of the current process, by checking in children to see if it's present. This is under the assumption that children will be filled, will need modifications.
Fixed spacing.