Writes to address passed to join
Switched from LinkedList of children to HashMap of children, this is under the assumption that it will be easier to handle, however, for the time being, the previous LinkedList code is still there, but commented out.
Changed halt() to only operate when the process calling it is the root process (pid == 0)
join() now acquires the status of the child process and writes it to the address passed into the method.